Continue fixing dtype errors on Windows
Kind of hard to do it in one go since CI is my only way to
reproduce the problem.

PiperOrigin-RevId: 801828915

+++ jax/_src/numpy/lax_numpy.py
@@ -7632,7 +7632,7 @@ def delete(
     obj = clip(where(obj < 0, obj + a.shape[axis], obj), 0, a.shape[axis])
     obj = sort(obj)
     obj -= arange(len(obj), dtype=obj.dtype)  # type: ignore
-    i = arange(a.shape[axis] - obj.size)
+    i = arange(a.shape[axis] - obj.size, dtype=obj.dtype)
     i += (i[None, :] >= obj[:, None]).sum(0)
     return a[(slice(None),) * axis + (i,)]
